// src/types/jobs/ValidateDiscountType.ts
export type ValidateDiscountType = {
    amount: number
    code: string
    createdDate: string
    endDate: string
    isValid: boolean
    name: string
    details: string
    id: number
    type: string
    startDate: string
    status: string
    discountedAmount: number
    totalAmount: number
}